api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.42.16",
    "fastapi[standard]>=0.124.4",
    "httpx>=0.28.1",
    "mypy-boto3>=1.42.3",
    "pydantic-settings>=2.12.0",
]
